+++ frontend/svelte/src/lib/constants/canisters.constants.ts
@@ -1,0 +1,2 @@
+// This minimum does not apply on testnets.
+export const NEW_CANISTER_MIN_T_CYCLES = 2;