+++ frontend/svelte/src/tests/lib/api/canisters.api.spec.ts
@@ -13,6 +13,7 @@ import {
   queryCanisterDetails,
   queryCanisters,
   topUpCanister,
+  updateSettings,
 } from "../../../lib/api/canisters.api";
 import {
   CREATE_CANISTER_MEMO,
@@ -27,7 +28,10 @@ import type { SubAccountArray } from "../../../lib/canisters/nns-dapp/nns-dapp.t
 import { CYCLES_MINTING_CANISTER_ID } from "../../../lib/constants/canister-ids.constants";
 import { mockSubAccount } from "../../mocks/accounts.store.mock";
 import { mockIdentity } from "../../mocks/auth.store.mock";
-import { mockCanisterDetails } from "../../mocks/canisters.mock";
+import {
+  mockCanisterDetails,
+  mockCanisterSettings,
+} from "../../mocks/canisters.mock";
 
 describe("canisters-api", () => {
   const mockNNSDappCanister = mock<NNSDappCanister>();
@@ -87,7 +91,38 @@ describe("canisters-api", () => {
     });
   });
 
-  describe("attachCanister", () => {
+  describe("updateSettings", () => {
+    afterEach(() => jest.clearAllMocks());
+
+    it("should call the ic management canister to update settings", async () => {
+      mockICManagementCanister.updateSettings.mockResolvedValue(undefined);
+      await updateSettings({
+        identity: mockIdentity,
+        canisterId: mockCanisterDetails.id,
+        settings: mockCanisterSettings,
+      });
+
+      expect(mockICManagementCanister.updateSettings).toBeCalled();
+    });
+
+    it("should call the ic management canister to update settings with partial settings", async () => {
+      const settings = {
+        controllers: [
+          "xlmdg-vkosz-ceopx-7wtgu-g3xmd-koiyc-awqaq-7modz-zf6r6-364rh-oqe",
+        ],
+      };
+      mockICManagementCanister.updateSettings.mockResolvedValue(undefined);
+      await updateSettings({
+        identity: mockIdentity,
+        canisterId: mockCanisterDetails.id,
+        settings,
+      });
+
+      expect(mockICManagementCanister.updateSettings).toBeCalled();
+    });
+  });
+
+  describe("detachCanister", () => {
     afterEach(() => jest.clearAllMocks());
 
     it("should call the nns dapp canister to detach the canister id", async () => {