+++ frontend/svelte/src/tests/routes/CanisterDetail.spec.ts
@@ -3,6 +3,8 @@
  */
 
 import { render, waitFor } from "@testing-library/svelte";
+import type { CanisterDetails } from "../../lib/canisters/ic-management/ic-management.canister.types";
+import { UserNotTheControllerError } from "../../lib/canisters/ic-management/ic-management.errors";
 import {
   getCanisterDetails,
   listCanisters,
@@ -14,13 +16,19 @@ import { mockCanister, mockCanisterDetails } from "../mocks/canisters.mock";
 import en from "../mocks/i18n.mock";
 import { mockRouteStoreSubscribe } from "../mocks/route.store.mock";
 
+const defaultReturn = Promise.resolve(mockCanisterDetails);
+let getCanisterDetailsReturn = defaultReturn;
+const setGetCanisterDetailReturn = (value: Promise<CanisterDetails>) =>
+  (getCanisterDetailsReturn = value);
+const resetGetCanisterDetailReturn = () =>
+  (getCanisterDetailsReturn = defaultReturn);
 jest.mock("../../lib/services/canisters.services", () => {
   return {
     listCanisters: jest.fn(),
     routePathCanisterId: () => mockCanister.canister_id.toText(),
     getCanisterDetails: jest
       .fn()
-      .mockImplementation(() => Promise.resolve(mockCanisterDetails)),
+      .mockImplementation(() => getCanisterDetailsReturn),
     getCanisterFromStore: () => mockCanister,
   };
 });
@@ -43,6 +51,18 @@ describe("CanisterDetail", () => {
     const { getByText } = render(CanisterDetail);
 
     expect(getByText(en.canister_detail.title)).toBeInTheDocument();
+  });
+
+  it("should render title once loaded", async () => {
+    canistersStore.setCanisters({ canisters: [mockCanister], certified: true });
+    const { container } = render(CanisterDetail);
+
+    const title = container.querySelector("h1");
+
+    await waitFor(() => expect(title).not.toBeNull());
+    expect((title as HTMLElement).textContent).toEqual(
+      mockCanister.canister_id.toText()
+    );
   });
 
   it("should fetch canisters from nns-dapp if store is not loaded yet", async () => {
@@ -78,5 +98,20 @@ describe("CanisterDetail", () => {
     );
     // Waiting for the one above is enough
     expect(queryByTestId("canister-controllers-card")).toBeInTheDocument();
+  });
+
+  it("should not render cards if user is not the controller", async () => {
+    setGetCanisterDetailReturn(Promise.reject(new UserNotTheControllerError()));
+    // Need to be the same that routePathCanisterId returns.
+    canistersStore.setCanisters({ canisters: [mockCanister], certified: true });
+    const { queryByTestId } = render(CanisterDetail);
+
+    await waitFor(() =>
+      expect(queryByTestId("canister-details-error-card")).toBeInTheDocument()
+    );
+    // Waiting for the one above is enough
+    expect(queryByTestId("canister-cycles-card")).not.toBeInTheDocument();
+    expect(queryByTestId("canister-controllers-card")).not.toBeInTheDocument();
+    resetGetCanisterDetailReturn();
   });
 });